+++ src/App.jsx
@@ -1,0 +1,11 @@
+import Header from './components/Header';
+
+function App() {
+    return (
+        <div>
+            <Header></Header>
+        </div>
+    )
+}
+
+export default App;
