--- src/assets/Component/Header-Item/style/CartIcon.css
@@ -1,16 +1,0 @@
-.cart-icon-container {
-    position: relative;
-    display: inline-block;
-  }
-  
-  .cart-count {
-    position: absolute;
-    top: 0;
-    right: 0;
-    background-color: red;
-    color: white;
-    border-radius: 50%;
-    padding: 2px 6px;
-    font-size: 12px;
-  }
-
