+++ src/styles/Header.css
@@ -1,0 +1,29 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans:ital,wght@0,600;1,600&display=swap');
+a {
+  color: inherit;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 10px;
+    background-color: white;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+    transition: box-shadow 0.3s ease;
+}
+
+.title {
+  font-family: "Noto Sans", sans-serif;
+  font-size: 24px;
+  font-weight: bold;
+  margin: 0;
+}
+  
+.icon {
+  font-size: 20px;
+  cursor: pointer;
+}
+
